# Conflicts:
#	src/main/java/org/example/pmas/model/User.java
#	src/main/java/org/example/pmas/repository/ProjectRepository.java
#	src/main/java/org/example/pmas/repository/SubProjectRepository.java
#	src/main/java/org/example/pmas/repository/TaskRepository.java
#	src/main/java/org/example/pmas/repository/UserRepository.java

+++ src/main/java/org/example/pmas/service/SubProjectService.java
@@ -24,10 +24,6 @@ public class SubProjectService {
         return subprojectRepository.readAll();
     }
 
-    public List<SubProject> getSubProjectsByProjectID(int projectId){
-        return subprojectRepository.getSubProjectsByProjectID(projectId);
-    }
-
     public SubProject readSelected(int id) {
         SubProject sub = subprojectRepository.readSelected(id);
         if(sub == null) {
@@ -48,6 +44,16 @@ public class SubProjectService {
     }
 
     public boolean delete(int id) {
+        if(!subprojectRepository.doesSubProjectExist(id)) {
+            throw new SubProjectNotFoundException(id);
+        }
         return subprojectRepository.delete(id);
+    }
+
+    public boolean updateSubProject(SubProject subproject) {
+        if(!subprojectRepository.doesSubProjectExist(subproject.getId())) {
+            throw new SubProjectNotFoundException(subproject.getId());
+        }
+        return subprojectRepository.update(subproject);
     }
 }
